Fixes in TL-B.md
`[A-z0-9_] letters` -> `[A-z0-9_] symbols`. Usually letters are characters of an alphabet, identifier characters are more often called as symbols. It is unusual to call digits as letters.

`After both separators, a undercore symbol` -> `After both separators, an undercore symbol`. It is better to use `an` before word starts with `u` letter.

Also typo fixes.